Reduce compile time
This PR reduces the huge compile time by a factor of two. I caught the
guilty template using the profiler features of `clang++`. Basically,
configure the project with

```
CXX=/usr/bin/clang++ CXXFLAGS="-ftime-trace" ./configure
```
will activate the profiler and `clang++` produces a json file per
compilation unit. The compiler instantiates massively the template `spdlog::pattern_formatter::handle_flag_` from the spdlog library. It is a well known issue (see for instance the issue https://github.com/gabime/spdlog/issues/1708). The solution consists in using the compiled version of the library with the flag `-DSPDLOG_COMPILED_LIB`.

Before this commit, the compile time for the complete project was around
3m30s on my laptop. After this commit, it is 1m50s.

The size of the runtime is 69M.